+++ CHANGELOG.md
@@ -15,6 +15,7 @@ We refer to [GitHub issues](https://github.com/eclipse/winery/issues) by using `
 * SELFSERVICE-Metadata is now additionally contained under the servicetemplates path in the CSAR
 * `csarName`, `version`, and `authors` are now contained in `data.xml` and `data.json`, too.
 * Enfore `LF` line endings in the repository
+* Add splitting functionality based on target labels
 
 ### Fixed
 * Bounary definitions can be browsed for exported operations again